backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
